ClassPilot-master-2/ClassPilot-master-2/requirements.txt
PyPDF2
streamlit
streamlit-reveal-slides
groq
langchain_groq
langchain
ollama
langchain_community
panda
numpy
pptx
gtts
pygame
